Fix major problem
When the Pi was booted during the time in which it should have been recording, it would record nonstop.  When the Pi was booted when it was not supposed to be recording, it would never record.  These modifications to the code with respect to the `time.get` reference should solve this problem.